requirements.txt
ShopifyAPI==12.3.0

